django/requirements.txt
asgiref==3.7.2
Django==4.1.2
django-filter==23.2
djangorestframework==3.14.0
python-json-logger==2.0.4
pytz==2023.3
sqlparse==0.4.4
psycopg2
